+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 
+### v1.0.1
+
+Improve error handling: filter for ips like "unknown" and "-"
+
+### v1.0.0
+
+Basic functionality